api/pyproject.toml
[project]
name = "mawster-api"
version = "0.1.0"
description = "Mawster API - MCOC Alliance Management"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "pyjwt",
    "python-multipart",
    "pydantic-settings",
    "sqlmodel",
    "aiomysql",
    "bleach",
    "sentry-sdk",
    "slowapi",
    "prometheus-fastapi-instrumentator",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "migrations/*", "scripts/*"]
concurrency = ["greenlet", "thread"]

[tool.ruff]

# Same as Black.
line-length = 100
indent-width = 4

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[dependency-groups]
migrate = [
    "alembic",
    "pymysql",
    "sqlmodel",
]
dev = [
    "aiosqlite>=0.22.1",
    "anyio>=4.12.1",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.0.0",
    "pytest-cov>=5.0.0",
    "pytest-xdist>=3.0.0",
    "vulture>=2.16",
    "faker>=24.0.0",
    "time-machine>=2.0.0",
    "icecream>=2.0.0",
    "beautifulsoup4>=4.0.0",
    "curl_cffi>=0.7.0",
    "lxml>=5.0.0",
    "Pillow>=10.0.0",
]
